Reduce repetitive code by instantiating TestClient in Injectedmockcontexttest

+++ uaa/src/test/java/org/cloudfoundry/identity/uaa/mock/InjectedMockContextTest.java
@@ -12,6 +12,12 @@
  *******************************************************************************/
 package org.cloudfoundry.identity.uaa.mock;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import org.apache.commons.codec.binary.Base64;
+import org.cloudfoundry.identity.uaa.oauth.token.TokenConstants;
+import org.cloudfoundry.identity.uaa.util.JsonUtils;
+import org.cloudfoundry.identity.uaa.util.SetServerNameRequestPostProcessor;
 import org.flywaydb.core.Flyway;
 import org.junit.AfterClass;
 import org.junit.Before;
@@ -23,12 +29,16 @@ import org.junit.runners.model.Statement;
 import org.springframework.restdocs.JUnitRestDocumentation;
 import org.springframework.security.web.FilterChainProxy;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.support.XmlWebApplicationContext;
 
 import static org.junit.Assume.assumeTrue;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
 import static org.springframework.restdocs.templates.TemplateFormats.markdown;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 public class InjectedMockContextTest implements Contextable {
 
@@ -39,14 +49,20 @@ public class InjectedMockContextTest implements Contextable {
     public JUnitRestDocumentation restDocumentation = new JUnitRestDocumentation("build/generated-snippets");
 
     private static XmlWebApplicationContext webApplicationContext;
-    private static MockMvc mockMvc;
+    private MockMvc mockMvc;
+    protected final TestClient testClient = new TestClient();
     private static volatile boolean mustDestroy = false;
 
-    public static XmlWebApplicationContext getWebApplicationContext() {
+    public static XmlWebApplicationContext getWebApplicationContext() throws Exception {
+        if (webApplicationContext == null) {
+            webApplicationContext = DefaultConfigurationTestSuite.setUpContext();
+            mustDestroy = true;
+        }
+
         return webApplicationContext;
     }
 
-    public static MockMvc getMockMvc() {
+    public MockMvc getMockMvc() {
         return mockMvc;
     }
 
@@ -55,14 +71,9 @@ public class InjectedMockContextTest implements Contextable {
     }
 
     @Before
-    public void initContextIfWeNeedIt() throws Exception {
-        if (getWebApplicationContext() == null) {
-            webApplicationContext = DefaultConfigurationTestSuite.setUpContext();
-            mustDestroy = true;
-        }
-
-        FilterChainProxy springSecurityFilterChain = webApplicationContext.getBean("springSecurityFilterChain", FilterChainProxy.class);
-        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
+    public void initMockMvc() throws Exception {
+        FilterChainProxy springSecurityFilterChain = getWebApplicationContext().getBean("springSecurityFilterChain", FilterChainProxy.class);
+        mockMvc = MockMvcBuilders.webAppContextSetup(getWebApplicationContext())
             .addFilter(springSecurityFilterChain)
             .apply(documentationConfiguration(this.restDocumentation)
                 .uris().withPort(80).and()
@@ -73,12 +84,11 @@ public class InjectedMockContextTest implements Contextable {
 
     @AfterClass
     public static void mustDestroy() throws Exception {
-        if (isMustDestroy()) {
+        if (isMustDestroy() && webApplicationContext != null) {
             webApplicationContext.getBean(Flyway.class).clean();
             webApplicationContext.destroy();
         }
         webApplicationContext = null;
-        mockMvc = null;
         mustDestroy = false;
     }
 
@@ -87,11 +97,69 @@ public class InjectedMockContextTest implements Contextable {
         webApplicationContext = context;
     }
 
+    public TestClient getTestClient() {
+        return testClient;
+    }
+
     public static class SkipWhenNotRunningInSuiteRule implements TestRule {
         @Override
         public Statement apply(Statement statement, Description description) {
             assumeTrue(UaaBaseSuite.shouldMockTestBeRun());
             return statement;
+        }
+    }
+
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    public static class OAuthToken {
+        @JsonProperty("access_token")
+        public String accessToken;
+
+        public OAuthToken() {
+        }
+    }
+
+    public class TestClient {
+        private TestClient() {
+        }
+
+        public String getClientCredentialsOAuthAccessToken(String username, String password, String scope) throws Exception {
+            return getClientCredentialsOAuthAccessToken(username, password, scope, null);
+        }
+
+        public String getClientCredentialsOAuthAccessToken(String username, String password, String scope, String subdomain)
+            throws Exception {
+            String basicDigestHeaderValue = "Basic "
+                + new String(Base64.encodeBase64((username + ":" + password).getBytes()));
+            MockHttpServletRequestBuilder oauthTokenPost = post("/oauth/token")
+                .header("Authorization", basicDigestHeaderValue)
+                .param("grant_type", "client_credentials")
+                .param("client_id", username)
+                .param(TokenConstants.REQUEST_TOKEN_FORMAT, TokenConstants.OPAQUE)
+                .param("scope", scope);
+            if (subdomain != null && !subdomain.equals(""))
+                oauthTokenPost.with(new SetServerNameRequestPostProcessor(subdomain + ".localhost"));
+            MvcResult result = mockMvc.perform(oauthTokenPost)
+                .andExpect(status().isOk())
+                .andReturn();
+            OAuthToken oauthToken = JsonUtils.readValue(result.getResponse().getContentAsString(), OAuthToken.class);
+            return oauthToken.accessToken;
+        }
+
+        public String getUserOAuthAccessToken(String clientId, String clientSecret, String username, String password, String scope)
+            throws Exception {
+            String basicDigestHeaderValue = "Basic "
+                + new String(Base64.encodeBase64((clientId + ":" + clientSecret).getBytes()));
+            MockHttpServletRequestBuilder oauthTokenPost = post("/oauth/token")
+                .header("Authorization", basicDigestHeaderValue)
+                .param("grant_type", "password")
+                .param("client_id", clientId)
+                .param("username", username)
+                .param("password", password)
+                .param(TokenConstants.REQUEST_TOKEN_FORMAT, TokenConstants.OPAQUE)
+                .param("scope", scope);
+            MvcResult result = mockMvc.perform(oauthTokenPost).andExpect(status().isOk()).andReturn();
+            OAuthToken oauthToken = JsonUtils.readValue(result.getResponse().getContentAsString(), OAuthToken.class);
+            return oauthToken.accessToken;
         }
     }
 }
